moved getBearer to own file
