+++ configuration-app-sample/src/public/js/app.js
@@ -1,7 +1,7 @@
 // import { port } from "../../../index.js";
 import { buildForm } from "./helpers/buildForm.js";
 import { formError } from "./helpers/formError.js";
-// console.log(port);
+import { getBearerToken } from "./helpers/getBearerToken.js";
 
 //build form
 buildForm();
@@ -9,26 +9,8 @@ buildForm();
 //get Access token
 const token = getBearerToken();
 
-function getBearerToken() {
-  try {
-    const query = decodeURI(window.location.search.substring(1));
-    const accessToken = query.split("access_token=")[1];
-    var host = decodeURI(window.location.origin);
-
-    // localStorage
-    localStorage.setItem("token", accessToken);
-    return JSON.parse(accessToken);
-  } catch (error) {
-    //link for token
-    const link = `${host}/index`;
-
-    let noTokenMessage = document.getElementById("mainArticle");
-    noTokenMessage.innerHTML = `
-      <h2 id="message" class="noToken"> Oops... <br> Looks like we dont have an access token.<br> Let's get one... <a class="noToken" href="${link}">oAuth</a></h2>
-    `;
-    return;
-  }
-}
+//get Bearer
+getBearerToken();
 
 let getResults = document.getElementById("submit");
 getResults.addEventListener("click", submitForm);
